pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastapi[standard]>=0.122.0",
    "phonenumbers>=9.0.19",
    "psycopg[binary]>=3.2.13",
    "pydantic-extra-types>=2.10.6",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "pre-commit>=4.5.0",
    "pytest>=9.0.1",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.6",
    "safety>=3.7.0",
]

[tool.ruff]
line-length = 100
extend-exclude = ['.venv', '__pycache__', '.html', '.log', '.ini']
indent-width = 4
[tool.ruff.format]
quote-style = 'single'
docstring-code-format = true
indent-style = 'space'
[tool.ruff.lint]
select = ['ALL']
ignore = ["E701", "D212", "D203", "ERA001", "ANN401", "PLR0913", "N802", "N803", "ARG001", "ARG002", "ARG004"]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ['D', 'S101', 'ANN201', 'ANN204', 'INP001', 'ANN001']
"__init__.py" = ['D']

[tool.black]
line-length = 100
target-version = ['py313']
extend-exclude = """
(
    .git
   | .html
   | .venv
   | .log
   | .ini
)
"""

[tool.bandit]
exclude_dirs = ['tests']
